requirements.txt
2captcha-python==1.1.2
pdfkit==1.0.0
selenium-wire==4.6.4
urllib3==1.26.7
bs4==0.0.1
google-api-python-client==2.52.0 
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
flask==2.1.2
gunicorn==20.1.0